Fix #77 changed grading to float (#134)

+++ backend/db_construct.sql
@@ -57,7 +57,7 @@ CREATE TABLE submissions (
 	submission_id INT GENERATED ALWAYS AS IDENTITY,
 	uid VARCHAR(255) NOT NULL,
 	project_id INT NOT NULL,
-	grading INTEGER CHECK (grading >= 0 AND grading <= 20),
+	grading FLOAT CHECK (grading >= 0 AND grading <= 20),
 	submission_time TIMESTAMP WITH TIME ZONE NOT NULL,
 	submission_path VARCHAR(50) NOT NULL,
 	submission_status submission_status NOT NULL,
